Create .gitignore
I haven't created my own gitignore before, but I will be creating a myriad of tests for my website to ensure that that there are no syntactic, semantic, or logic errors in my server side. When testing these my target folder will be changing each time I run the tests and thusly need to be ignored when creating PRs and such.